Moved permissions and kill to unstable
